+++ src/App.css
@@ -46,7 +46,7 @@ body {
   padding: 1.5rem;
   width: auto;
   background-color: #fff;
-  border: solid 2px rgb(122, 121, 121);
+  border: solid 2px rgba(122, 121, 121, 0.322);
   /* border-radius: 12px; */
 
 }
@@ -73,8 +73,9 @@ body {
 .interactions {
   display: flex;
   position: absolute;
+  align-items: center;
   justify-content: flex-end;
-  right: 0;
+  right: 10px;
   bottom: 20px;
 }
 
@@ -84,8 +85,8 @@ body {
 .remove-btn {
   cursor: pointer;
   font-size: 1.8rem;
-  align-self: flex-end;
   align-self: center;
+  margin: 1rem 0 0 1rem;
 }
 .remove-btn:hover {
   color: rgb(226, 25, 25);
@@ -133,7 +134,7 @@ body {
   background-color: #fff;
   padding: 1.5rem;
   margin: 0.7rem 30px;
-  border: solid 2px rgb(122, 121, 121);
+  border: solid 2px rgba(122, 121, 121, 0.322);
   /* border-radius: 10px; */
   font-family: Verdana, Geneva, Tahoma, sans-serif;
 }
@@ -158,7 +159,7 @@ body {
 .wishlist:hover {
   color: red;
 }
-@media screen and (max-width: 925px) {
+@media screen and (max-width:1055px) {
   .cart-card {
     flex-direction: column;
     justify-content: center;
@@ -170,14 +171,13 @@ body {
   }
   .quant-btn {
     margin-top: 1rem;
-    margin-left: -2rem;
+    /* margin-left: -2rem; */
   }
 }
-@media screen and (max-width: 768px) {
+@media screen and (max-width: 900px) {
   .filter {
     font-size: 1rem;
   }
-
   .place-btn {
     align-self: center;
   }
@@ -193,6 +193,9 @@ body {
     flex-direction: column;
     justify-content: space-evenly;
     align-items: center;
+  }
+  .card-header{
+    text-align: center;
   }
   .card-main-info {
     justify-content: space-evenly;
@@ -216,5 +219,6 @@ body {
     position: static;
     justify-content: center;
     flex-direction: column;
+    margin-right: 0 ;
   }
 }
